pyproject.toml
[project]
name = "gen-story"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.54.0",
    "pydantic-settings>=2.12.0",
    "streamlit>=1.52.1",
]

[tool.uv.sources]
src = { path = "src" }
